PDFAnalyzer/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
qdrant-client
langchain-community
langchain-text-splitters
langchain-openai
pypdf
tiktoken
python-multipart
gradio
requests
